fix(upgrade): extract logic and fix bug for validate form login
