+++ src/components/pages/LoginPage/index.tsx
@@ -5,16 +5,16 @@ import { Grid } from "@mui/material";
 import { Form, ErrorsCard } from "../../molecules";
 import { PrimaryText } from "../../atoms";
 import { handleLogin } from "../../../integrations";
+import { validateInput } from "../../../utils";
 
-interface LoginPageProps extends RouteComponentProps {
-  id: string;
-}
+interface LoginPageProps extends RouteComponentProps {}
 
-export const LoginPage: React.FC<LoginPageProps> = ({ match, history }) => {
+export const LoginPage: React.FC<LoginPageProps> = ({ history }) => {
   const [username, setUsername] = useState<string>("");
   const [password, setPassword] = useState<string>("");
   const [usernameErrors, setUsernameErrors] = useState<string[]>([]);
   const [passwordErrors, setPasswordErrors] = useState<string[]>([]);
+
   const handleOnChangeUser = (e: { target: { value: string } }) => {
     setUsername(e.target.value);
   };
@@ -22,39 +22,9 @@ export const LoginPage: React.FC<LoginPageProps> = ({ match, history }) => {
     setPassword(e.target.value);
   };
 
-  const validateInput = async (type: string, input: string) => {
-    const localErrors: string[] = [];
-    let inputType: string = type === "password" ? "password" : "username";
-    let value: string;
-    const expression: RegExp = /[^A-Za-z0-9]+/g;
-
-    if (input.length < 6) {
-      localErrors.push(`Your ${inputType} must be at least 6 characters`);
-    } else {
-      if (input.trim() !== input) {
-        localErrors.push(`Your ${inputType} should not contain spaces`);
-      } else {
-        if (!!input.match(expression)) {
-          localErrors.push(
-            `Your ${inputType} should not contain special characters`
-          );
-        }
-      }
-    }
-    try {
-      if (localErrors.length === 0) {
-        value = input;
-      } else {
-        value = "";
-      }
-    } catch {
-      localErrors.push("Sorry. Something went wrong. Try again.");
-      value = "";
-    }
-    return { hasErrors: localErrors, input: value };
-  };
-
   const handleLoginForm = () => {
+    setUsernameErrors([]);
+    setPasswordErrors([]);
     validateInput("username", username).then((r) => {
       if (r) {
         const { hasErrors, input } = r;
